Dealing with FD false positives
We may be affected by false positives in terms of unreachability by the FD in use. In this case, instruct the unreachable member to shut down (it still gets the message, unless there really is a partition) and move back into the WaitingForMembers state.

Also handle added and removed members properly at other stages than WaitingForMembers